Questioner/requirements.txt
google-generativeai>=0.8.0
pydantic>=2.5.0
